[refactor] 패키징 수정
패키징 수정

Relate to : 25
