+++ app/src/main/java/meaning/morning/presentation/data/HomeCardData.kt
@@ -1,2 +1,2 @@
-package meaning.morning
+package meaning.morning.presentation.data
